chore: Update database and hangfire configurations
This commit updates the database and hangfire configurations in the FSH.Starter project. In the database.json file, the connection string has been modified to include additional security settings such as encryption and trust server certificate. Similarly, in the hangfire.json file, the storage provider has been changed to "mssql" and the connection string has been updated with the same additional security settings. These changes are necessary to enhance the security of the application's database and hangfire storage.

Note: Please remove this note before committing the changes.