feat: Add Fundraisers property to AccountDto and additional properties to CampaignDto
These code changes add the Fundraisers property to the AccountDto class in the FSH.Starter project. This property is a list of FundraiserDto objects and enhances the functionality of the AccountDto.

Additionally, the CampaignDto class is updated with the TargetAmount property, as well as the Donations, CampaignStudents, and Configurations properties. These changes improve the data structure and capabilities of the CampaignDto.