feat: Include related entities in query handlers
This commit includes changes to the query handlers in the Campaign module of the FSH.Starter project. The GetCampaignStudentByIdQueryHandler, GetDonationStudentByIdQueryHandler, GetAllAccountsQueryHandler, GetAllCampaignStudentsQueryHandler, GetAllDonationsQueryHandler, and GetAccountByIdQueryHandler classes have been modified to include related entities in their queries. This improves the efficiency and accuracy of retrieving campaign, donation, student, and account data.